refactor `bus_mapping_tmp` and `bus_mapping_tmp_convert` (#18)
* move `bus_mapping_tmp` and `bus_mapping_tmp_convert` to witness.rs

* impl From<&bus_mapping::circuit_input_builder::ExecStep> for ExecutionResult

* impl From<&bus_mapping::bytecode::Bytecode> for bus_mapping_tmp::Bytecode

* cargo fmt

* merge

* re-direct

* cargo fmt

* unwrap

* unwrap

* clean up
