+++ zkevm-circuits/src/evm_circuit/execution/and.rs
@@ -1,9 +1,6 @@
 use crate::{
     evm_circuit::{
-        execution::{
-            bus_mapping_tmp::{Block, Call, ExecStep, Transaction},
-            ExecutionGadget,
-        },
+        execution::ExecutionGadget,
         step::ExecutionResult,
         table::{FixedTableTag, Lookup},
         util::{
@@ -13,6 +10,7 @@ use crate::{
             },
             Word,
         },
+        witness::{Block, Call, ExecStep, Transaction},
     },
     util::Expr,
 };
@@ -102,8 +100,8 @@ impl<F: FieldExt> ExecutionGadget<F> for AndGadget<F> {
 #[cfg(test)]
 mod test {
     use crate::evm_circuit::{
-        bus_mapping_tmp_convert,
         test::{rand_word, run_test_circuit_complete_fixed_table},
+        witness,
     };
     use bus_mapping::{bytecode, eth_types::Word};
 
@@ -123,10 +121,7 @@ mod test {
             XOR
             STOP
         };
-        let block =
-            bus_mapping_tmp_convert::build_block_from_trace_code_at_start(
-                &bytecode,
-            );
+        let block = witness::build_block_from_trace_code_at_start(&bytecode);
 
         assert_eq!(run_test_circuit_complete_fixed_table(block), Ok(()));
     }
