+++ zkevm-circuits/src/evm_circuit/execution/jumpdest.rs
@@ -1,9 +1,6 @@
 use crate::{
     evm_circuit::{
-        execution::{
-            bus_mapping_tmp::{Block, Call, ExecStep, Transaction},
-            ExecutionGadget,
-        },
+        execution::ExecutionGadget,
         step::ExecutionResult,
         util::{
             common_gadget::SameContextGadget,
@@ -11,6 +8,7 @@ use crate::{
                 ConstraintBuilder, StateTransition, Transition::Delta,
             },
         },
+        witness::{Block, Call, ExecStep, Transaction},
     },
     util::Expr,
 };
@@ -55,7 +53,7 @@ impl<F: FieldExt> ExecutionGadget<F> for JumpdestGadget<F> {
 #[cfg(test)]
 mod test {
     use crate::evm_circuit::{
-        bus_mapping_tmp_convert, test::run_test_circuit_incomplete_fixed_table,
+        test::run_test_circuit_incomplete_fixed_table, witness,
     };
     use bus_mapping::bytecode;
 
@@ -65,10 +63,7 @@ mod test {
             JUMPDEST
             STOP
         };
-        let block =
-            bus_mapping_tmp_convert::build_block_from_trace_code_at_start(
-                &bytecode,
-            );
+        let block = witness::build_block_from_trace_code_at_start(&bytecode);
         assert_eq!(run_test_circuit_incomplete_fixed_table(block), Ok(()));
     }
