+++ zkevm-circuits/src/evm_circuit/execution/memory.rs
@@ -1,9 +1,6 @@
 use crate::{
     evm_circuit::{
-        execution::{
-            bus_mapping_tmp::{Block, Call, ExecStep, Transaction},
-            ExecutionGadget,
-        },
+        execution::ExecutionGadget,
         param::MAX_GAS_SIZE_IN_BYTES,
         step::ExecutionResult,
         util::{
@@ -17,6 +14,7 @@ use crate::{
             memory_gadget::MemoryExpansionGadget,
             select, MemoryAddress, Word,
         },
+        witness::{Block, Call, ExecStep, Transaction},
     },
     util::Expr,
 };
@@ -202,8 +200,8 @@ impl<F: FieldExt> ExecutionGadget<F> for MemoryGadget<F> {
 #[cfg(test)]
 mod test {
     use crate::evm_circuit::{
-        bus_mapping_tmp_convert,
         test::{rand_word, run_test_circuit_incomplete_fixed_table},
+        witness,
     };
     use bus_mapping::{
         bytecode,
@@ -318,7 +316,7 @@ mod test {
             bytecodes: vec![bytecode],
         };*/
         //let block  =
-        // bus_mapping_tmp_convert::build_block_from_trace_code_at_start(&
+        // witness::build_block_from_trace_code_at_start(&
         // bytecode);
         let gas = Gas(gas_cost + 100_000); // add extra gas for the pushes
         let mut block_trace =
@@ -340,8 +338,7 @@ mod test {
 
         //println!("old block is {:#?}", builder.block);
 
-        let block =
-            bus_mapping_tmp_convert::block_convert(&bytecode, &builder.block);
+        let block = witness::block_convert(&bytecode, &builder.block);
 
         //println!("block is {:#?}", block);
