+++ zkevm-circuits/src/evm_circuit/execution/pc.rs
@@ -1,9 +1,6 @@
 use crate::{
     evm_circuit::{
-        execution::{
-            bus_mapping_tmp::{Block, Call, ExecStep, Transaction},
-            ExecutionGadget,
-        },
+        execution::ExecutionGadget,
         step::ExecutionResult,
         util::{
             common_gadget::SameContextGadget,
@@ -12,6 +9,7 @@ use crate::{
             },
             from_bytes, RandomLinearCombination,
         },
+        witness::{Block, Call, ExecStep, Transaction},
     },
     util::Expr,
 };
@@ -83,7 +81,7 @@ impl<F: FieldExt> ExecutionGadget<F> for PcGadget<F> {
 #[cfg(test)]
 mod test {
     use crate::evm_circuit::{
-        bus_mapping_tmp_convert, test::run_test_circuit_incomplete_fixed_table,
+        test::run_test_circuit_incomplete_fixed_table, witness,
     };
     use bus_mapping::bytecode;
 
@@ -94,10 +92,7 @@ mod test {
             PC
             STOP
         };
-        let block =
-            bus_mapping_tmp_convert::build_block_from_trace_code_at_start(
-                &bytecode,
-            );
+        let block = witness::build_block_from_trace_code_at_start(&bytecode);
 
         assert_eq!(run_test_circuit_incomplete_fixed_table(block), Ok(()));
     }
