+++ zkevm-circuits/src/evm_circuit/execution/push.rs
@@ -1,9 +1,6 @@
 use crate::{
     evm_circuit::{
-        execution::{
-            bus_mapping_tmp::{Block, Call, ExecStep, Transaction},
-            ExecutionGadget,
-        },
+        execution::ExecutionGadget,
         step::ExecutionResult,
         util::{
             common_gadget::SameContextGadget,
@@ -12,6 +9,7 @@ use crate::{
             },
             sum, Cell, Word,
         },
+        witness::{Block, Call, ExecStep, Transaction},
     },
     util::Expr,
 };
@@ -154,8 +152,8 @@ impl<F: FieldExt> ExecutionGadget<F> for PushGadget<F> {
 #[cfg(test)]
 mod test {
     use crate::evm_circuit::{
-        bus_mapping_tmp_convert,
         test::{rand_bytes, run_test_circuit_incomplete_fixed_table},
+        witness,
     };
     use bus_mapping::{bytecode, evm::OpcodeId};
 
@@ -172,10 +170,7 @@ mod test {
             bytecode.write(*b);
         }
         bytecode.write_op(OpcodeId::STOP);
-        let block =
-            bus_mapping_tmp_convert::build_block_from_trace_code_at_start(
-                &bytecode,
-            );
+        let block = witness::build_block_from_trace_code_at_start(&bytecode);
         assert_eq!(run_test_circuit_incomplete_fixed_table(block), Ok(()));
     }
