+++ zkevm-circuits/src/evm_circuit/execution/swap.rs
@@ -1,9 +1,6 @@
 use crate::{
     evm_circuit::{
-        execution::{
-            bus_mapping_tmp::{Block, Call, ExecStep, Transaction},
-            ExecutionGadget,
-        },
+        execution::ExecutionGadget,
         step::ExecutionResult,
         util::{
             common_gadget::SameContextGadget,
@@ -12,6 +9,7 @@ use crate::{
             },
             Cell, Word,
         },
+        witness::{Block, Call, ExecStep, Transaction},
     },
     util::Expr,
 };
@@ -95,8 +93,8 @@ impl<F: FieldExt> ExecutionGadget<F> for SwapGadget<F> {
 #[cfg(test)]
 mod test {
     use crate::evm_circuit::{
-        bus_mapping_tmp_convert,
         test::{rand_word, run_test_circuit_incomplete_fixed_table},
+        witness,
     };
     use bus_mapping::{bytecode, eth_types::Word, evm::OpcodeId};
 
@@ -115,10 +113,7 @@ mod test {
             .write_op(opcode)
             STOP
         });
-        let block =
-            bus_mapping_tmp_convert::build_block_from_trace_code_at_start(
-                &bytecode,
-            );
+        let block = witness::build_block_from_trace_code_at_start(&bytecode);
         assert_eq!(run_test_circuit_incomplete_fixed_table(block), Ok(()));
     }
